add manual tests with servo and led

+++ src/functions/functions.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 extern void displayBTCheck();
-extern void ledFunctionality(uint8_t milliseconds);
-extern void servoFunctionality(uint8_t milliseconds);
+extern void ledFunctionality(uint16_t milliseconds);
+extern void servoFunctionality(uint16_t milliseconds);
 
 #endif
